+++ src/main/java/domain/card/Card.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public class Card {
     private final CardType cardType;
