+++ src/main/java/domain/card/CardValue.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum CardValue {
     ACE(1),
