+++ src/main/java/domain/participant/Money.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 public class Money {
