+++ src/main/java/domain/participant/Player.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.participant;
+
+import domain.card.Cards;
 
 public class Player extends Participant {
