+++ src/main/java/domain/result/DealerScore.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import java.util.List;
