+++ src/main/java/domain/result/PlayerScore.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public class PlayerScore {
