+++ src/test/java/domain/CardDeckTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.CardDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
