+++ src/test/java/domain/CardTest.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.CardType;
+import domain.card.CardValue;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
