+++ src/test/java/domain/JudgeTest.java
@@ -1,5 +1,12 @@
 package domain;
 
+import domain.card.*;
+import domain.participant.Dealer;
+import domain.participant.Money;
+import domain.participant.Name;
+import domain.participant.Player;
+import domain.result.Judge;
+import domain.result.PlayerScore;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
