+++ src/test/java/domain/MoneyTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.participant.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
