+++ src/test/java/domain/NameTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.participant.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
