+++ src/test/java/domain/PlayerTest.java
@@ -6,6 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import java.util.ArrayList;
 import java.util.List;
 
+import domain.card.*;
+import domain.participant.Money;
+import domain.participant.Name;
+import domain.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
