+++ src/test/java/domain/ProfitTest.java
@@ -1,5 +1,13 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.CardType;
+import domain.card.CardValue;
+import domain.card.Cards;
+import domain.participant.Money;
+import domain.participant.Name;
+import domain.participant.Player;
+import domain.result.Profit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
